+++ index.html
@@ -95,10 +95,18 @@
 			</div>
 		</div>
 
-		<input id="inputGroup" type="number" min = "1" value="1" step="1" inputmode="numeric">
-		<button class="genButton">Generate Group</button>
-		<div class="tableShow">
-			<h1>Showwwwwwwwwwwww</h1>
+		<div class="container2">
+			<h1 class="result-banner">Result</h1>
+			<div>
+				<input id="inputGroup" type="number" min = "1" value="3" step="1" inputmode="numeric">
+				<button class="genButton">Generate Group</button>
+			</div>
+		</div>
+
+		<div class="container3">
+			<div class="tableShow">
+				<h1 id = "result"></h1>
+			</div>
 		</div>
 		<script src="./js/main.js" defer type="module"></script>
